mod generation command
