+++ packages/cli/src/utilities.ts
@@ -1,5 +1,6 @@
 import { Command } from "commander";
 import fs from "node:fs";
+import path from "node:path";
 
 import { FHIR_VERSION, R4, R4B } from "@iguhealth/fhir-types/versions";
 
@@ -23,6 +24,20 @@ export namespace conversion {
       }
     }
   }
+}
+
+export function getAllFiles(directory: string): string[] {
+  const files: string[] = [];
+  const filesInDirectory = fs.readdirSync(directory);
+  for (const file of filesInDirectory) {
+    const absolute = path.join(directory, file);
+    if (fs.statSync(absolute).isDirectory()) {
+      files.push(...getAllFiles(absolute));
+    } else {
+      files.push(absolute);
+    }
+  }
+  return files;
 }
 
 export namespace dataCommand {
